create contactFormLogic.js file and replace duplicated logic with it. import navbarScrollResponsive and menuOpenClose to empathic-building page
